Address PR feedback: move function to data module and fix precedence logic
- Move computeActiveAreaMediaStates to src/data/media-player.ts
- Make areaId required parameter and rename function for clarity
- Fix precedence logic: playing state takes precedence over paused
- Update tests and remove redundant validation
